+++ src/main/java/com/bitbank/controllers/BankAccountTransactionsController.java
@@ -1,6 +1,6 @@
 package com.bitbank.controllers;
 
-import jakarta.validation.Valid;
+import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -26,6 +27,7 @@ import com.bitbank.services.BankAccountTransactionsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/api/v1/bank-account-transactions")
@@ -53,6 +55,16 @@ public class BankAccountTransactionsController {
 		return new ResponseEntity<>(bankAccountTransactionsService.index(page, perPage, orderBy, orderWay),
 				new HttpHeaders(), HttpStatus.OK);
 
+	}
+
+	/**
+	 * Show. Single transaction
+	 */
+	@Operation(security = { @SecurityRequirement(name = "bearer-key") })
+	@GetMapping("/{id}")
+	@PreAuthorize("hasRole('CUSTOMER')")
+	public ResponseEntity<Optional<BankAccountTransactionsEntity>> show(@PathVariable Long id) {
+		return new ResponseEntity<>(bankAccountTransactionsService.show(id), new HttpHeaders(), HttpStatus.OK);
 	}
 
 	/**